Added house API calls and databases
Kaleo, Timothy, and Brandon were able to figure out how to connect to the houses API and populate the appropriate table. The rest should follow in the same way. Next step should be to update the house API calls with lat and long, and concurrently we need to get the crime data web scraper up. We also need to see if taking data out of the databases is trivial (fingers crossed).