Add files via upload
Modularized the clustering algorithm so that all the clustering for anything whatsoever is done by cluster_data.py, all we need to do is call it from a different file. Then I added a main file that automatically calls for the location data and the price data table to empty themselves, cluster the data, and repopulate themselves with those centroids. The make_table.py file is only there to make it easier for the programmer to create new tables as needed or to view tables already created.