Add files via upload
Starts the final.py file, which should put everything that's needed into the database. In this commit, we begin by sorting by zipcode as the PK, and storing the percentage of each zipcode from each cluster. It then "assigns" a cluster to each zipcode in accordance to which cluster has the most houses from each zipcode.